refactor: create PDAL package
